# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Aelu API
  version: 1.0.0
  description: >
    REST API for the Aelu learning system. Supports adaptive drill sessions
    (via WebSocket), spaced-repetition progress tracking, graded reading, media
    exposure, extensive listening, vocabulary encounter logging, xAPI/Caliper
    learning-record export, Common Cartridge export, GDPR data portability and
    erasure, JWT-based mobile auth, offline sync, Stripe payments, and push
    notifications.
servers:
  - url: https://aelu.app
security:
  - cookieAuth: []
  - bearerAuth: []
...
paths:
...
  /api/sync/push:
    post:
      tags: [Sync]
      summary: Push batched offline actions to the server
      operationId: syncPush
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [actions]
              properties:
                actions:
                  type: array
                  items:
                    type: object
                    properties:
                      type:
                        type: string
                        enum: [drill_result, vocab_encounter, media_watched]
                      data:
                        type: object
                      timestamp:
                        type: string
                        format: date-time
      responses:
        "200":
          description: Sync result
          content:
            application/json:
              schema:
                type: object
                properties:
                  processed:
                    type: integer
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        index:
                          type: integer
                        error:
                          type: string
                  total:
                    type: integer
  /api/sync/pull:
    get:
      tags: [Sync]
      summary: Pull new progress and sessions since a timestamp
      operationId: syncPull
      parameters:
        - name: since
          in: query
          schema:
            type: string
            format: date-time
            default: "1970-01-01T00:00:00"
          description: Return data updated after this timestamp
      responses:
        "200":
          description: Updated data
          content:
            application/json:
              schema:
                type: object
                properties:
                  progress:
                    type: array
                    items:
                      type: object
                  sessions:
                    type: array
                    items:
                      type: object
                  since:
                    type: string
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: Flask session cookie set on login
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from /api/auth/token
...